Financial Connections: for new playground app, did some renaming.
