+++ Example/FinancialConnections Example/FinancialConnections Example/Playground/PlaygroundViewController.swift
@@ -14,7 +14,7 @@ final class PlaygroundViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         // create
-        let hostingController = UIHostingController(rootView: PlaygroundMainView())
+        let hostingController = UIHostingController(rootView: PlaygroundView())
 
         // add to subview
         view.addSubview(hostingController.view)
